Use double instead of float for midi frequency

+++ include/iimavlib/midi/MidiTypes.h
@@ -59,12 +59,12 @@ inline uint8_t name_to_note(const note_name name)
     return static_cast<uint8_t>(name);
 }
 
-inline float note_to_freq(const uint8_t note)
+inline double note_to_freq(const uint8_t note)
 {
-    return std::pow(2.f, (note - 69) / 12.f) * 440.f;
+    return std::pow(2.0, (note - 69) / 12.0) * 440.0;
 }
 
-inline float name_to_freq(const note_name name)
+inline double name_to_freq(const note_name name)
 {
     return note_to_freq(name_to_note(name));
 }
